improve readability of index file
